helper_requirements.txt
numpy==1.16.1
h5py==2.10.0
cython==0.29.15
pandas==0.25.3
scikit-learn==0.22.2.post1
scipy==1.4.1
PyYAML
open3d-python==0.3.0.0
tensorflow-gpu==1.11
laspy==1.7.0